* master:
  Add a test that parallels an Operator plugin test in network repo (#633)
  updated ens-sync-script to work with fastchain (#615)
  Sponsorship fix: don't let _reduceStakeBy revert when over-slashing (#632)
  doc: update build and publish images description
  doc: publish instructions
  release(network-contracts): v6.1.0
  release(config): @streamr/config 4.0.5
  ETH-585 docker images build and test in CI (#616)
  Eth 596 subgraph fixeschanges (#631)
  fix: merge hazard from flagWithMetadata PR
  cleanup: reorder setParam function, uint256 -> uint in Policies
  feat: make slashing fraction configurable (#624)
  rename flagWithMetadata -> flag, and remove the previous implementation (#630)
  release(config): @streamr/config 4.0.4
  streamr 1.0 mumbai deployment addresses update

# Conflicts:
#	package-lock.json
#	packages/network-subgraphs/subgraph.yaml

+++ packages/network-contracts/contracts/OperatorTokenomics/testcontracts/TestAllocationPolicy.sol
@@ -20,7 +20,7 @@ contract TestAllocationPolicy is IAllocationPolicy, Sponsorship {
         assembly {data.slot := storagePosition} // solhint-disable-line no-inline-assembly
     }
 
-    function setParam(uint256 testCase) external {
+    function setParam(uint testCase) external {
         if (testCase == 1) {
             require(false, "test_setParam");
         } else if (testCase == 2) {
@@ -59,7 +59,7 @@ contract TestAllocationPolicy is IAllocationPolicy, Sponsorship {
     }
 
     /** Horizon means how long time the (unallocated) funds are going to still last */
-    function getInsolvencyTimestamp() public override pure returns (uint256 horizonSeconds) {
+    function getInsolvencyTimestamp() public override pure returns (uint horizonSeconds) {
         return 2**255; // indefinitely solvent
     }
 
@@ -79,7 +79,7 @@ contract TestAllocationPolicy is IAllocationPolicy, Sponsorship {
     }
 
     /** Calculate the cumulative earnings per unit (full token stake) right now */
-    function getCumulativeEarnings() internal view returns(uint256) {
+    function getCumulativeEarnings() internal view returns(uint) {
     }
 
     function onSponsor(address, uint) external {
@@ -88,6 +88,6 @@ contract TestAllocationPolicy is IAllocationPolicy, Sponsorship {
     function getEarningsWei(address) public view returns (uint earningsWei) {
     }
 
-    function calculatePenaltyOnStake(address) external view returns (uint256 stake) {
+    function calculatePenaltyOnStake(address) external view returns (uint stake) {
     }
 }
